requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-filter==22.1
django-jsonify==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==17.6.0
importlib-metadata==6.0.0
Markdown==3.4.1
passlib==1.7.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
zipp==3.15.0
